colocate UI types of the `but-core::diff` module
