+++ crates/but-core/src/diff/mod.rs
@@ -3,3 +3,6 @@ pub use commit::commit_changes;
 
 mod worktree;
 pub use worktree::worktree_changes;
+
+/// conversion functions for use in the UI
+pub mod ui;
