+++ crates/but-core/tests/core/diff/mod.rs
@@ -1,2 +1,3 @@
 mod commit_changes;
+mod ui;
 pub(crate) mod worktree_changes;
