--- crates/gitbutler-tauri/tests/tauri/main.rs
@@ -1,1 +1,0 @@
-mod diff;
